Detect when devices are added and removed
Add a connection listener to handle LOST and CONNECTED status.
This requires the Chromecasts to be connected when the app is
started. If the app in running while a Chromecast is rebooted,
It will attempt to dynamically remove and re-add the device.
Also, there is a global REBOOT_VOLUME that can be set at the
top of the file. This value will be used to adjust the volume
after reboot since Chromecasts always reboot to maximum volume.
This requires https://github.com/balloob/pychromecast/pull/305