Overhaul playback and add auto-play for files
Introduce catt process management and auto-play feature. Since catt doesn't
exit when playing files, we need to manage the processes we've started.
This uses os.kill to send SIGINT to the catt process and waits to avoid
zombie processes. All catt processes are killed on clean shutdown. This
also implmenets auto-play for files. The directory of the file played is
searched for other files and they are played in alphabetical order.