Add files via upload
The main problem seems to be that QPoint does not accept floating point values. It must have at one time otherwise this app wouldn't have worked for so long. For now i have wrapped the floating point numbers in an int() function. It seems to fix the issue as far as what I have tested. I'm not really familiar with Qt, but the values I converted to int seem to correspond to text size and the volume changer. Do let me know if I have messed up the logic of your app here.